[Refactor] Consolidate `typings/` into `@types/`, and API types
