--- src/@types/api/pokerogue-api-types.ts
@@ -1,4 +1,0 @@
-export interface TitleStatsResponse {
-  playerCount: number;
-  battleCount: number;
-}
