--- src/@types/api/pokerogue-save-data-api.ts
@@ -1,8 +1,0 @@
-import type { SessionSaveData, SystemSaveData } from "#types/save-data";
-
-export interface UpdateAllSavedataRequest {
-  system: SystemSaveData;
-  session: SessionSaveData;
-  sessionSlotId: number;
-  clientSessionId: string;
-}
