+++ src/@types/i18next.d.ts
