--- src/@types/session-save-migrator.ts
@@ -1,6 +1,0 @@
-import type { SessionSaveData } from "./save-data";
-
-export interface SessionSaveMigrator {
-  version: string;
-  migrate: (data: SessionSaveData) => void;
-}
