--- src/@types/settings-save-migrator.ts
@@ -1,5 +1,0 @@
-export interface SettingsSaveMigrator {
-  version: string;
-  // biome-ignore lint/complexity/noBannedTypes: TODO - refactor settings
-  migrate: (data: Object) => void;
-}
