+++ src/plugins/api/pokerogue-savedata-api.ts
@@ -2,7 +2,7 @@ import { ApiBase } from "#api/api-base";
 import { PokerogueSessionSavedataApi } from "#api/pokerogue-session-savedata-api";
 import { PokerogueSystemSavedataApi } from "#api/pokerogue-system-savedata-api";
 import { MAX_INT_ATTR_VALUE } from "#app/constants";
-import type { UpdateAllSavedataRequest } from "#types/api/pokerogue-save-data-api";
+import type { UpdateAllSavedataRequest } from "#types/api";
 
 /**
  * A wrapper for PokéRogue savedata API requests.
