+++ src/plugins/api/pokerogue-session-savedata-api.ts
@@ -6,7 +6,7 @@ import type {
   GetSessionSavedataRequest,
   NewClearSessionSavedataRequest,
   UpdateSessionSavedataRequest,
-} from "#types/api/pokerogue-session-save-data-api";
+} from "#types/api";
 import type { SessionSaveData } from "#types/save-data";
 
 /**
