+++ src/system/version-migration/version-converter.ts
@@ -2,9 +2,7 @@
 
 import { version } from "#package.json";
 import type { SessionSaveData, SystemSaveData } from "#types/save-data";
-import type { SessionSaveMigrator } from "#types/session-save-migrator";
-import type { SettingsSaveMigrator } from "#types/settings-save-migrator";
-import type { SystemSaveMigrator } from "#types/system-save-migrator";
+import type { SessionSaveMigrator, SettingsSaveMigrator, SystemSaveMigrator } from "#types/save-migrators";
 import { compareVersions } from "compare-versions";
 
 /*
