+++ src/system/version-migration/versions/v1_10_0.ts
@@ -3,7 +3,7 @@ import type { MoveId } from "#enums/move-id";
 import type { MoveResult } from "#enums/move-result";
 import { MoveUseMode } from "#enums/move-use-mode";
 import type { SessionSaveData } from "#types/save-data";
-import type { SessionSaveMigrator } from "#types/session-save-migrator";
+import type { SessionSaveMigrator } from "#types/save-migrators";
 import type { TurnMove } from "#types/turn-move";
 
 /** Prior signature of `TurnMove`; used to ensure parity */
