+++ src/system/version-migration/versions/v1_8_3.ts
@@ -1,7 +1,7 @@
 import { DexAttr } from "#enums/dex-attr";
 import { SpeciesId } from "#enums/species-id";
 import type { SystemSaveData } from "#types/save-data";
-import type { SystemSaveMigrator } from "#types/system-save-migrator";
+import type { SystemSaveMigrator } from "#types/save-migrators";
 import { getPokemonSpecies } from "#utils/pokemon-utils";
 
 /**
