+++ test/plugins/api/pokerogue-account-api.test.ts
@@ -2,7 +2,7 @@ import { PokerogueAccountApi } from "#api/pokerogue-account-api";
 import { SESSION_ID_COOKIE_NAME } from "#app/constants";
 import { initServerForApiTests } from "#test/test-utils/test-file-initialization";
 import { getApiBaseUrl } from "#test/test-utils/test-utils";
-import type { AccountInfoResponse } from "#types/api/pokerogue-account-api";
+import type { AccountInfoResponse } from "#types/api";
 import * as CookieUtils from "#utils/cookies";
 import * as cookies from "#utils/cookies";
 import { HttpResponse, http } from "msw";
