+++ test/plugins/api/pokerogue-api.test.ts
@@ -1,7 +1,7 @@
 import { pokerogueApi } from "#api/pokerogue-api";
 import { initServerForApiTests } from "#test/test-utils/test-file-initialization";
 import { getApiBaseUrl } from "#test/test-utils/test-utils";
-import type { TitleStatsResponse } from "#types/api/pokerogue-api-types";
+import type { TitleStatsResponse } from "#types/api";
 import { HttpResponse, http } from "msw";
 import type { SetupServerApi } from "msw/node";
 import { afterEach, beforeAll, beforeEach, describe, expect, it, vi } from "vitest";
