+++ test/plugins/api/pokerogue-savedata-api.test.ts
@@ -1,7 +1,7 @@
 import { PokerogueSavedataApi } from "#api/pokerogue-savedata-api";
 import { initServerForApiTests } from "#test/test-utils/test-file-initialization";
 import { getApiBaseUrl } from "#test/test-utils/test-utils";
-import type { UpdateAllSavedataRequest } from "#types/api/pokerogue-save-data-api";
+import type { UpdateAllSavedataRequest } from "#types/api";
 import { HttpResponse, http } from "msw";
 import type { SetupServerApi } from "msw/node";
 import { afterEach, beforeAll, beforeEach, describe, expect, it, vi } from "vitest";
