+++ test/plugins/api/pokerogue-system-savedata-api.test.ts
@@ -6,7 +6,7 @@ import type {
   UpdateSystemSavedataRequest,
   VerifySystemSavedataRequest,
   VerifySystemSavedataResponse,
-} from "#types/api/pokerogue-system-save-data-api";
+} from "#types/api";
 import type { SystemSaveData } from "#types/save-data";
 import { HttpResponse, http } from "msw";
 import type { SetupServerApi } from "msw/node";
